Converted web project to same react template as notification gateway
