+++ Source/Alerts/Web/.babelrc
@@ -1,8 +1,3 @@
 {
-    "presets": ["@babel/preset-env", "@babel/preset-react"],
-    "plugins": [
-        "@babel/plugin-syntax-dynamic-import",
-        "@babel/plugin-proposal-class-properties",
-        ["@babel/proposal-decorators", { "legacy": true }]
-    ]
+  "extends": "./node_modules/@dolittle/build.react/.babelrc"
 }
