--- Source/Alerts/Web/.env
@@ -1,2 +1,0 @@
-UPSTREAM_API_URL=http://www.dev.cbsrc.org
-API_URL=http://localhost:3001
