--- Source/Alerts/Web/.env.development
@@ -1,1 +1,0 @@
-API_URL=http://localhost:3000
