--- Source/Alerts/Web/.env.nightwatch
@@ -1,1 +1,0 @@
-API_URL=http://localhost:9001
