--- Source/Alerts/Web/.eslintignore
@@ -1,8 +1,0 @@
-node_modules
-public
-*.html
-*.log
-*.json
-*.md
-*.yaml
-*.ejs
