--- Source/Alerts/Web/.eslintrc
@@ -1,87 +1,0 @@
-{
-    "extends": ["eslint:recommended", "plugin:react/recommended"],
-    "plugins": ["import", "react"],
-    "parser": "babel-eslint",
-    "parserOptions": {
-        "ecmaVersion": 6,
-        "sourceType": "module",
-        "ecmaFeatures": {
-            "jsx": true,
-            "experimentalObjectRestSpread": true
-        }
-    },
-    "env": {
-        "es6": true,
-        "browser": true,
-        "jest": true,
-        "node": true,
-        "jquery": true,
-        "mocha": true
-    },
-    "rules": {
-        "eqeqeq": ["error", "smart"],
-        "no-extra-boolean-cast": 0,
-        "quotes": ["error", "single"],
-        "curly": ["error", "all"],
-        "no-console": [
-            "warning",
-            {
-                "allow": ["warning"]
-            }
-        ],
-        "no-debugger": 2,
-        "semi": [1, "always", { "omitLastInOneLineBlock": true }],
-        "semi-spacing": [2, { "before": false, "after": true }],
-        "no-trailing-spaces": 2,
-        "no-else-return": 2,
-        "no-extra-bind": 2,
-        "no-implicit-coercion": 2,
-        "no-useless-call": 2,
-        "no-return-assign": 2,
-        "eol-last": 2,
-        "no-unused-vars": 1,
-        "no-extra-semi": 1,
-        "no-underscore-dangle": 1,
-        "no-lone-blocks": 2,
-        "one-var": ["error", "never"],
-        "no-var": 2,
-        "no-alert": 2,
-        "no-constant-condition": [
-            "error",
-            {
-                "checkLoops": false
-            }
-        ],
-        "jsx-quotes": 1,
-        "react/display-name": 0,
-        "react/jsx-uses-react": 1,
-        "react/jsx-no-undef": 2,
-        "react/jsx-wrap-multilines": 2,
-        "react/no-string-refs": 0,
-        "react/jsx-boolean-value": 1,
-        "react/jsx-closing-bracket-location": 0,
-        "react/jsx-curly-spacing": 1,
-        "react/jsx-indent-props": 0,
-        "react/jsx-key": 1,
-        "react/jsx-max-props-per-line": 0,
-        "react/jsx-no-bind": 0,
-        "react/jsx-no-duplicate-props": 1,
-        "react/jsx-no-literals": 0,
-        "react/jsx-pascal-case": 1,
-        "react/jsx-sort-prop-types": 0,
-        "react/jsx-sort-props": 0,
-        "react/jsx-uses-vars": 1,
-        "react/no-danger": 1,
-        "react/no-did-mount-set-state": 1,
-        "react/no-did-update-set-state": 1,
-        "react/no-direct-mutation-state": 1,
-        "react/no-multi-comp": 1,
-        "react/no-set-state": 0,
-        "react/no-unknown-property": 1,
-        "react/prefer-es6-class": 1,
-        "react/prop-types": 0,
-        "react/react-in-jsx-scope": 1,
-        "react/self-closing-comp": 1,
-        "react/sort-comp": 1
-    }
-}
