--- Source/Alerts/Web/.prettierrc
@@ -1,6 +1,0 @@
-{
-    "printWidth": 100,
-    "tabWidth": 4,
-    "singleQuote": true,
-    "trailingComma": "es5"
-}
