--- Source/Alerts/Web/.sassrc
@@ -1,5 +1,0 @@
-{
-  "includePaths": [
-    "node_modules"
-  ]
-}
