+++ Source/Alerts/Web/Environments/development.env
@@ -1,0 +1,1 @@
+API_BASE_URL=http://localhost:5010
