+++ Source/Alerts/Web/Environments/production.env
@@ -1,0 +1,1 @@
+API_BASE_URL=/alerts
