+++ Source/Alerts/Web/index.ejs
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <meta charset="utf-8">
+        <title><%- htmlWebpackPlugin.options.metadata.title %></title>
+        <meta name="viewport" content="width=device-width, initial-scale=1">
+    
+        <base href="<%- htmlWebpackPlugin.options.metadata.baseUrl %>">
+        <!-- imported CSS are concatenated and added automatically -->
+      </head>
+      <body>
+        <div id="root"></div>
+      </body>
+</html>
