+++ Source/Alerts/Web/jsconfig.json
@@ -1,0 +1,5 @@
+{
+    "compilerOptions": {
+        "experimentalDecorators": true
+    }
+}
