+++ Source/Alerts/Web/karma.conf.js
@@ -1,0 +1,2 @@
+"use strict";
+module.exports = require("@dolittle/build/karma.conf.js");
