+++ Source/Alerts/Web/package.json
@@ -1,71 +1,38 @@
 {
-    "name": "@ifrc-cbs/alerts",
-    "publishConfig": {
-        "access": "public"
-    },
-    "version": "0.0.4",
-    "description": "CBS Alerts frontend React component",
-    "scripts": {
-        "start": "npm-run-all -l -p node:server parcel:dev",
-        "start:local": "npm-run-all -l -p local:server parcel:dev",
-        "dev": "npm run parcel:dev",
-        "node:server": "cross-env PORT=3001 NODE_ENV=development cbs-node-server",
-        "local:server": "cross-env PORT=3001 TESTING_MODE=true NODE_ENV=development cbs-node-server",
-        "test": "cross-env HEADLESS_MODE=1 nightwatch",
-        "export": "babel --config-file=./.babel/.babelrc --verbose --out-dir=es src && cp -R src/assets es",
-        "clean": "rimraf es",
-        "parcel:dev": "cross-env NODE_ENV=development parcel ./public/index.html",
-        "prepare": "npm run clean && npm run export",
-        "eslint": "eslint src",
-        "build": "npm run eslint && npm run prepare"
-    },
-    "main": "es/exports.js",
-    "author": "",
-    "license": "MIT",
-    "files": [
-        "src",
-        "es"
-    ],
-    "dependencies": {
-        "@babel/runtime": "^7.2.0",
-        "@ifrc-cbs/common-react-ui": "file:../../Navigation/Web.Common",
-        "@ifrc-cbs/reporting": "0.0.4",
-        "bootstrap-sass": "^3.4.0",
-        "@dolittle/commands": "^1.0.12",
-        "@dolittle/queries": "^1.0.14",
-        "dotenv": "^6.2.0",
-        "evergreen-ui": "^4.9.0",
-        "prop-types": "^15.6.2",
-        "react": "^16.7.0",
-        "redux": "^4.0.1",
-        "redux-logger": "^3.0.6",
-        "redux-saga": "^0.16.2"
-    },
-    "devDependencies": {
-        "@babel/cli": "^7.2.3",
-        "@babel/core": "^7.2.2",
-        "@babel/plugin-external-helpers": "^7.2.0",
-        "@babel/plugin-proposal-class-properties": "^7.2.3",
-        "@babel/plugin-proposal-decorators": "^7.2.3",
-        "@babel/plugin-syntax-dynamic-import": "^7.2.0",
-        "@babel/plugin-transform-runtime": "^7.2.0",
-        "@babel/preset-react": "^7.0.0",
-        "@ifrc-cbs/nodejs-server": "file:../../Navigation/Web.NodeJS",
-        "babel-eslint": "^10.0.1",
-        "chromedriver": "^2.45.0",
-        "cross-env": "^5.2.0",
-        "eslint": "^5.12.0",
-        "eslint-plugin-import": "^2.14.0",
-        "eslint-plugin-react": "^7.12.3",
-        "nightwatch": "^1.0.18",
-        "npm-run-all": "^4.1.5",
-        "parcel-bundler": "^1.11.0",
-        "prettier": "^1.15.3",
-        "react-dom": "^16.7.0",
-        "react-helmet": "^5.2.0",
-        "react-redux": "^6.0.0",
-        "react-router-dom": "^4.3.1",
-        "rimraf": "^2.6.3",
-        "sass": "^1.16.0"
-    }
+  "name": "web",
+  "version": "1.0.0",
+  "description": "",
+  "main": "karma.conf.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "webpack --mode production --config webpack.config.prod.js",
+    "start": "webpack-dev-server --mode development --open --watch --progress --hot",
+    "run": "cd .. && dolittle run"
+  },
+  "author": "",
+  "license": "ISC",
+  "devDependencies": {
+    "@dolittle/build": "^6.2.3",
+    "@dolittle/build.react": "^1.0.6",
+    "@dolittle/styles": "^2.9.5",
+    "dotenv-webpack": "^1.7.0",
+    "webpack": "^4.26.0",
+    "webpack-cli": "^3.1.2",
+    "webpack-dev-server": "^3.1.10",
+    "webpack-notifier": "^1.7.0"
+  },
+  "dependencies": {
+    "@dolittle/commands": "^1.0.10",
+    "@dolittle/core": "^1.0.4",
+    "@dolittle/queries": "^1.0.12",
+    "evergreen-ui": "^4.9.0",
+    "formik": "^1.5.1",
+    "react": "^16.7.0",
+    "react-dom": "^16.7.0",
+    "react-redux": "^7.0.1",
+    "react-router-dom": "^5.0.0",
+    "redux": "^4.0.1",
+    "redux-logger": "^3.0.6",
+    "redux-saga": "^1.0.2"
+  }
 }
