--- Source/Alerts/Web/public/index.html
@@ -1,37 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <meta charset="utf-8" />
-        <title>Alerts | CBS</title>
-        <link
-            href="https://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css"
-            rel="stylesheet"
-            integrity="sha384-wvfXpqpZZVQGK6TAh5PVlGOfQNHSoD2xbE+QkPxCAFlNEevoEH3Sl0sibVcOQVnN"
-            crossorigin="anonymous"
-        />
-    </head>
-    <body>
-        <div id="app"></div>
-        <!--
-  This HTML file is a template.
-  If you open it directly in the browser, you will see an empty page.
-
-  It is only used for developing the Alerts module independently.
-
-  IMPORTANT:
-  Changes from this HTML file will NOT be included in the final HTML template
-
-  To begin the development, run `npm start`.
--->
-        <script>
-            if (window.console) {
-                console.log(
-                    'You are running the Alerts module development server.\n' +
-                        'Please note that if you are making changes in the index.html template file located in this module ' +
-                        'those changes will NOT be included in the main template.'
-                );
-            }
-        </script>
-        <script src="../index.js"></script>
-    </body>
-</html>
