+++ Source/Alerts/Web/src/index.js
@@ -1,0 +1,26 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import { BrowserRouter, Route } from 'react-router-dom';
+import { Provider } from 'react-redux';
+
+import App from './components/App';
+import store from './store';
+
+import './assets/main.scss';
+
+const routes = [
+    {
+        component: App,
+        route: '/alerts/',
+        exact: false,
+    }
+];
+
+ReactDOM.render(
+    <Provider store={store}>
+        <BrowserRouter>
+            <Route path="/" exact component={App} />
+        </BrowserRouter>
+    </Provider>,
+    document.getElementById('root')
+);
