+++ Source/Alerts/Web/wallaby.js
@@ -1,0 +1,1 @@
+module.exports = require('@dolittle/build.react/wallaby.js')('Features', () => {});
