+++ Source/Alerts/Web/webpack.config.prod.js
@@ -1,0 +1,2 @@
+process.env.DOLITTLE_WEBPACK_BASE_URL = '/alerts/'
+module.exports = require('./webpack.config.js');
